refactor(html pipe): rename image unwrap library
rename the hero image unwrap library to be more generic

fix #338
