+++ src/defaults/html.pipe.js
@@ -31,7 +31,7 @@ const validate = require('../utils/validate');
 const { cache, uncached } = require('../html/shared-cache');
 const embeds = require('../html/find-embeds');
 const parseFrontmatter = require('../html/parse-frontmatter');
-const unwrapHeroImages = require('../html/unwrap-hero-images.js');
+const unwrapHeroImages = require('../html/unwrap-sole-images');
 const rewriteLinks = require('../html/static-asset-links');
 const tovdom = require('../html/html-to-vdom');
 const tohtml = require('../html/stringify-response');
